refactor ConvertPermsToString

+++ src/commons/util/util.cpp
@@ -1058,7 +1058,7 @@ bool ParseInt32(const std::string& str, int32_t *out)
 
 bool  ConvertPermsToString(uint64_t perms, uint8_t total_perms_count, string& permsList) {
     ostringstream s;
-    for (uint8_t i = total_perms_count - 1; i >= 0; i--) {
+    for (int i = total_perms_count - 1; i >= 0; i--) {
         uint64_t p = (perms & ( 1 << i )) > 0 ? 1 :0;
         s << p;
     }
